refactor(parquet): split the ALP implementation by concern
One file held the scaling arithmetic, the parameter search and the page
format. Each now has its own file and owns the constants it uses:

- alp_scaling.go turns a float into the integer ALP stores, and back.
- alp_params.go picks the exponent and factor for a vector.
- alp_encoding.go writes and reads the bytes of a page.

Three simplifications came out of the split. The exponent and factor pair
had three representations, so the shortlist and its counting map now use
alpEncodingParams directly, which drops the uint32 packing and unpacking.
alpBitWidth wrapped bits.Len64 without adding anything. The two Type
methods held the same switch, which is now alpParquetType.

A vector header size replaces the constant covering only its first three
fields, so the decoder reads the whole header before validating it and
advances once.

Also document the encoder and decoder methods whose contract is not
obvious from the name, and cover the widest packing a page can carry:
1024 values whose differences need all 64 bits.
